pyproject.toml
[tool.poetry]
name = "crkp"
version = "0.1.0"
description = ""
authors = ["Prayag Chatha <pgchatha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.2.1"
lightning = "^2.5.0.post0"
pandas = "^2.2.3"
scipy = "^1.14.1"
jupyter = "^1.1.1"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
pytest = "^8.3.4"
hydra-core = "^1.3.2"
wandb = "^0.19.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
